+++ treasury/src/config.rs
@@ -14,13 +14,12 @@ pub const GROUP_CONTRACT: &str =
 pub const USDC_ADDRESS: &str =
     "CBIELTK6YBZJU5UP2WWQEUCYKLPU6AUNZ2BQ4WWFEIE3USCIHMXQDAMA";
 
+/// Replace with real mainnet contract IDs before deploying.
 #[cfg(all(not(test), feature = "mainnet"))]
-pub const GROUP_CONTRACT: &str =
-    "CA_MAINNET...";
+pub const GROUP_CONTRACT: &str = "CA_MAINNET...";
 
 #[cfg(all(not(test), feature = "mainnet"))]
-pub const USDC_ADDRESS: &str =
-    "CA_MAINNET...";
+pub const USDC_ADDRESS: &str = "CA_MAINNET...";
 
 #[cfg(not(any(test, feature = "testnet", feature = "mainnet")))]
 compile_error!("You must enable either 'testnet' or 'mainnet' feature");
