+++ treasury/src/tests/mod.rs
@@ -1,0 +1,8 @@
+//! Treasury contract test suite.
+//!
+//! Split by feature: treasury creation, fund rounds, releases, group balance.
+
+mod create_treasury_test;
+mod fund_round_test;
+mod release_test;
+mod group_balance_test;
